src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-networking>=5.0.0,<5.1.0
